Moved components to a different file
